requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.0.0
Django==6.0.1
google-generativeai==0.8.5
django-allauth==65.13.1
filelock==3.20.3
fsspec==2026.1.0
huggingface-hub==0.36.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numpy==1.26.4
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==4.57.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.6.3
zope.interface==8.2
google-generativeai==0.8.5

psycopg2-binary==2.9.10
pgvector==0.3.6
